Adesso il main è sincronizzato con il test

+++ project/Entity.cpp
@@ -4,7 +4,7 @@
 
 #include "Entity.h"
 
-int Entity::getHp() const {
+float Entity::getHp() const {
     return hp;
 }
